Refactor Game class methods for easier testing
Some methods were combined, #play_game, #restart, and #input_checker
because they did not need to be different methods when the only thing
that changed between the old #play_game and #restart methods was the
method name being passed into #input_checker and the #puts when the user
inputs a no response.

Some of the code in various methods were extracted and made into their
own methods:
- #create_board
- #verify_confirmation_input
- #verify_number_input methods.
This is to make testing easier.

The #check_position_input method was also renamed to #place_mark because
the input verification was moved to its own method:
- #verify_number_input.

The #winner and #win_positions methods were also changed so that
the #win_positions method returns either a nil or an array of winning
positions, so #winner can be more DRY.