Refactor Board class methods for testing
Methods that deal with printing the board were changed into private
methods and a new #player_symbol_input was created to handle getting
the player's input.

The #assign_symbol method was changed so that it only deals with
assigning the symbols and any verification of the symbols is dealt with
in other new methods like #symbol and #verify_symbol.

The #inc_round_and_return_symbol method was renamed to #player_symbol.